Fix issue with meta key getting stuck
There was a race condition when switching from within a vim window which
caused the meta key to become stuck, and every keystroke thereafter to
likely be interpreted by i3 as a command.

The race condition was eliminated by taking a completely new approach
which doesn't rely on libxdo to simulate keystrokes. Instead, vim is now
expected to be launched in client/server mode, and the `--remote-expr`
feature is now used to directly call the method defined in our vim
plugin.

Instructions in the README have been updated to reflect this. This is a
breaking change.

Prior to release, I plan to update the command line arguments handling
to be robust (ie, the vim binary option should require an option name
rather than being positional) as to not cause a future breaking change.